requirements.txt
django
djangorestframework
ipython
ipdb